+++ packages/lee_video_video_player/lib/lee_video_video_player.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/official_video_player_kernel.dart';